[PATCH 005/125] [MAKEFILE] Properly label build/docker artifacts as
 .SECONDARY

Some objects such as build/docker/bin/peer end up being classified
as .INTERMEDIATE objects because of the way we use wildcarding.  The
side effect is that these artifacts may be implicitly deleted in
many/most circumstances, the result of which is that they then
need to be recompiled _every_ time rather than taking advantage
of build-caching.  We can fix this by marking them explicitly
as .SECONDARY artifacts

Fixes FAB-243

Change-Id: If45b0f5c7863d26a061831373187895069573fbb
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>