[PATCH 017/125] Nominating Srinivasan Muralidharan to Fabric
 Maintainer

Srinivasan Muralidharan (Murali) is the leader on chaincode infrastructure
and actively participates in code reviews across components. Murali has been
instrumental in working with developers from LSEG, DTCC, DAH, and others to
bring about changes not only on chaincodes but eventing, SDK, and storage.
He is now leading the effort to stand up the skeleton for the new architecture
so that we can hang the rest on. Therefore, I'd like to nominate Murali to
maintainer to help keeping Fabric code at high quality. Thanks.

Change-Id: Ic0455d52581465fb3e7c36ef7104c280be133b60
Signed-off-by: Binh Q. Nguyen <binhn@us.ibm.com>