[PATCH 047/125] Check for duplicates when adding peers to the chain
Prior to this fix, it was possible to add duplicate peers
to the chain.  This fix now adds check to the addPeer method
and throws a DuplicatePeer error if a peer with the same URL is
already a member of the chain.  Also added a test for adding multiple
valid peers to the chain

Fixes FAB-264

Change-Id: I575200325bf8e8aedc72e1f41419099689c3a6fc
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>