[PATCH 055/125] [NODESDK] Do not use sudo/global in Makefile
The current SDK build relies on "npm install -g" for some of its
tooling.  This requires sudo in some cases, which can be problematic.
This also means we are spraying binaries around a system which may not
be appreciated in some environments.

The global install is not strictly necessary as long as we can be
intelligent about the paths of our prerequisite tooling. We therefore
refactor the SDK build to use local installs with explicit path
management.  We also clean up the build process a bit so we add some
build-avoidance in places where it is reasonable to do so.

Change-Id: I802d2b3894e9b0a7cea07a1ac577fa1061634ebb
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>