[PATCH 088/125] FAB-394 Chaincode log level cannot be changed
These changes make sure that the logging.chaincode setting from
core.yaml is used to set the chaincode logging level every time
the chaincode container starts. The peer reads in the value and
passes it to the chaincode via an environment variable as it
does for similar values.

Fix Issue FAB-394

Change-Id: Ic4b7228c57fc673a97dbfafc1b086ad04c41c05c
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>