[PATCH 094/125] Use ccenv docker image rather than baseimage
We want baseimage to only be used as a mechanism to assist
the build, not referenced at runtime.  Therefore, we
convert most instances of "baseimage" to use "ccenv"
(chaincode environment) since this is what they really want
anyway.

Note 1: A few instances (such as the UTXO example) did not have
an obvious solution so I left those to be sorted out by
the respective subsystem owners.

Note 2: The GOLANG chaincode type is currently a little suboptimal
in that the fabric-ccenv image loads the full fabric.git src
at build time and then the runtime re-injects it.  In the future
the build will only inject the golang shim and the runtime will
only inject the chaincode.  For now, this is somewhat inefficient
but ultimately still results in correct behavior.

Change-Id: I7f025bfff147fcca0f29f1656c7f63adc7a17173
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>