[PATCH 115/125] FAB-425 - improve initial experience
Improve the documentation to make it easier to find things,
and to make the standalone sample development environment
more prominent.

Fixed broken link to SystemChaincodes/noop.md so that it is
integrated into the documentation.

Refactored the standalone doc to read better and to limit
line length to 80.

Added release notes.
Changed to starter-kit

Change-Id: I8d5f0eb62a6ac2e5e338c3c9e486c4f952f9e0e2
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>